refactor: split App.jsx into components
