+++ monitor/src/components/layout/SleepCountdown.jsx
@@ -1,0 +1,23 @@
+import React, { useState, useEffect } from 'react'
+
+export default function SleepCountdown({ sleepUntil }) {
+  const [remaining, setRemaining] = useState('')
+  
+  useEffect(() => {
+    const update = () => {
+      const diff = sleepUntil - Date.now()
+      if (diff <= 0) {
+        setRemaining('Starting...')
+        return
+      }
+      const mins = Math.floor(diff / 60000)
+      const secs = Math.floor((diff % 60000) / 1000)
+      setRemaining(`${mins}m ${secs}s`)
+    }
+    update()
+    const interval = setInterval(update, 1000)
+    return () => clearInterval(interval)
+  }, [sleepUntil])
+  
+  return <span className="text-sm font-mono text-blue-600">{remaining}</span>
+}
